requirements.txt
pdf2image>=1.16.0
opencv-python>=4.8.0
pytesseract>=0.3.10
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
